feat: scaffold initial application structure and implement core UI components including navigation.
