+++ index.html
@@ -1,15 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
 
-  <!DOCTYPE html>
-  <html lang="en">
-    <head>
-      <meta charset="UTF-8" />
-      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <title>Website Animation Build Prompt</title>
-    </head>
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Galagama Gems - Exquisite Gemstones & Jewelry</title>
+  <meta name="description"
+    content="Discover exquisite gemstones and handcrafted jewelry at Galagama Gems. Shop our collection of sapphires, moonstones, and rubies." />
+  <meta property="og:title" content="Galagama Gems - Exquisite Gemstones & Jewelry" />
+  <meta property="og:description"
+    content="Discover exquisite gemstones and handcrafted jewelry at Galagama Gems. Shop our collection of sapphires, moonstones, and rubies." />
+  <meta property="og:type" content="website" />
+  <meta property="og:image" content="/og-image.jpg" />
+</head>
 
-    <body>
-      <div id="root"></div>
-      <script type="module" src="/src/main.tsx"></script>
-    </body>
-  </html>
-  
+<body>
+  <div id="root"></div>
+  <script type="module" src="/src/main.tsx"></script>
+</body>
+
+</html>
