+++ package.json
@@ -43,21 +43,40 @@
             "react-dom": "^18.3.1",
             "react-hook-form": "^7.55.0",
             "react-resizable-panels": "^2.1.7",
+            "react-router-dom": "^7.9.6",
             "recharts": "^2.15.2",
             "sonner": "^2.0.3",
             "tailwind-merge": "*",
             "vaul": "^1.1.2"
       },
       "devDependencies": {
+            "@eslint/js": "^9.39.1",
+            "@testing-library/jest-dom": "^6.9.1",
+            "@testing-library/react": "^16.3.0",
             "@types/node": "^20.10.0",
+            "@types/react": "^19.2.6",
+            "@types/react-dom": "^19.2.3",
+            "@typescript-eslint/eslint-plugin": "^8.47.0",
+            "@typescript-eslint/parser": "^8.47.0",
             "@vitejs/plugin-react-swc": "^3.10.2",
+            "eslint": "^9.39.1",
+            "eslint-plugin-react": "^7.37.5",
+            "eslint-plugin-react-hooks": "^7.0.1",
+            "eslint-plugin-react-refresh": "^0.4.24",
             "gh-pages": "^6.3.0",
-            "vite": "^5.4.11"
+            "globals": "^16.5.0",
+            "jsdom": "^27.2.0",
+            "prettier": "^3.6.2",
+            "typescript": "^5.9.3",
+            "typescript-eslint": "^8.47.0",
+            "vite": "^5.4.11",
+            "vitest": "^4.0.13"
       },
       "scripts": {
             "dev": "npx vite",
             "build": "npx vite build",
             "predeploy": "npm run build",
-            "deploy": "gh-pages -d dist"
+            "deploy": "gh-pages -d dist",
+            "test": "vitest"
       }
 }
