+++ src/components/CTAButton.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import { motion } from 'motion/react';
+import { useNavigate } from 'react-router-dom';
 
 interface CTAButtonProps {
   children: React.ReactNode;
@@ -9,12 +10,13 @@ interface CTAButtonProps {
 
 export function CTAButton({ children, variant, href }: CTAButtonProps) {
   const [isHovered, setIsHovered] = useState(false);
+  const navigate = useNavigate();
 
   const isPrimary = variant === 'primary';
 
   const handleClick = () => {
     if (href) {
-      window.location.hash = href;
+      navigate(href);
     }
   };
 
@@ -29,14 +31,14 @@ export function CTAButton({ children, variant, href }: CTAButtonProps) {
       transition={{ duration: 0.2, ease: "easeOut" }}
       className={`
         relative px-8 py-3 rounded-lg transition-all duration-300
-        ${isPrimary 
-          ? 'bg-gradient-to-r from-amber-400 to-yellow-600 text-black' 
+        ${isPrimary
+          ? 'bg-gradient-to-r from-amber-400 to-yellow-600 text-black'
           : 'bg-white/10 backdrop-blur-sm text-white border border-white/30'
         }
       `}
       style={{
-        boxShadow: isHovered 
-          ? '0 10px 30px -5px rgba(251, 191, 36, 0.5)' 
+        boxShadow: isHovered
+          ? '0 10px 30px -5px rgba(251, 191, 36, 0.5)'
           : '0 4px 6px -1px rgba(0, 0, 0, 0.1)'
       }}
     >
