+++ src/components/Home.tsx
@@ -1,0 +1,233 @@
+import { useState, useEffect } from 'react';
+import { motion, useScroll, useTransform } from 'motion/react';
+import { Navigation } from './Navigation';
+import { HeroContent } from './HeroContent';
+import { Footer } from './Footer';
+import { useCart } from '../context/CartContext';
+
+interface HomeProps {
+    setToastMessage: (msg: string) => void;
+    setShowToast: (show: boolean) => void;
+}
+
+export function Home({ setToastMessage, setShowToast }: HomeProps) {
+    const [isLoaded, setIsLoaded] = useState(false);
+    const { scrollY } = useScroll();
+    const { addToCart } = useCart();
+
+    // Parallax effect - background moves slower than foreground
+    const backgroundY = useTransform(scrollY, [0, 500], [0, 150]);
+
+    useEffect(() => {
+        setIsLoaded(true);
+    }, []);
+
+    return (
+        <div className="min-h-screen">
+            {/* Header Section with Parallax */}
+            <div className="relative h-screen overflow-hidden">
+                {/* Background Image with Ken Burns Effect and Parallax */}
+                <motion.div
+                    style={{ y: backgroundY }}
+                    className="absolute inset-0 w-full h-full"
+                >
+                    <motion.div
+                        initial={{ scale: 1, opacity: 0 }}
+                        animate={{ scale: 1.1, opacity: 1 }}
+                        transition={{
+                            scale: {
+                                duration: 20,
+                                ease: "linear",
+                                repeat: Infinity,
+                                repeatType: "reverse"
+                            },
+                            opacity: {
+                                duration: 1.2,
+                                ease: "easeOut"
+                            }
+                        }}
+                        className="w-full h-full"
+                    >
+                        <motion.div
+                            className="w-full h-full bg-cover bg-center"
+                            initial={{ filter: "brightness(0.7) contrast(1)" }}
+                            animate={{
+                                filter: [
+                                    "brightness(0.7) contrast(1)",
+                                    "brightness(0.8) contrast(1.05)",
+                                    "brightness(0.7) contrast(1)"
+                                ]
+                            }}
+                            transition={{
+                                duration: 8,
+                                ease: "easeInOut",
+                                repeat: Infinity,
+                                repeatType: "reverse"
+                            }}
+                            style={{
+                                backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.5)), url('https://images.unsplash.com/photo-1515562141207-7a88fb7ce338?w=1920&q=80')`
+                            }}
+                        />
+                    </motion.div>
+                </motion.div>
+
+                {/* Overlay gradient for better text readability */}
+                <div className="absolute inset-0 bg-gradient-to-b from-black/30 via-transparent to-black/50" />
+
+                {/* Content */}
+                <div className="relative z-10 h-full flex flex-col">
+                    <Navigation isLoaded={isLoaded} />
+                    <HeroContent isLoaded={isLoaded} />
+                </div>
+            </div>
+
+            {/* Featured Products Section */}
+            <div className="relative py-20 px-8" style={{
+                background: 'linear-gradient(180deg, #0f172a 0%, #1e293b 100%)',
+            }}>
+                {/* Decorative overlay */}
+                <div className="absolute inset-0 opacity-5" style={{
+                    backgroundImage: 'radial-gradient(circle at 20% 50%, rgba(251, 191, 36, 0.3) 0%, transparent 50%), radial-gradient(circle at 80% 80%, rgba(251, 191, 36, 0.2) 0%, transparent 50%)'
+                }} />
+
+                <div className="max-w-6xl mx-auto relative z-10">
+                    <motion.h2
+                        initial={{ opacity: 0, y: 20 }}
+                        whileInView={{ opacity: 1, y: 0 }}
+                        viewport={{ once: true }}
+                        transition={{ duration: 0.6 }}
+                        className="text-center mb-12"
+                        style={{
+                            fontSize: '2.5rem',
+                            fontWeight: '700',
+                            background: 'linear-gradient(to right, rgb(251, 191, 36), rgb(202, 138, 4))',
+                            WebkitBackgroundClip: 'text',
+                            WebkitTextFillColor: 'transparent',
+                            backgroundClip: 'text'
+                        }}
+                    >
+                        Our Featured Products
+                    </motion.h2>
+                    <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
+                        {[
+                            {
+                                id: 1,
+                                name: 'Sapphire Elegance Ring',
+                                category: 'ring',
+                                metal: '18K Gold',
+                                price: 850,
+                                description: 'Exquisite blue sapphire set in lustrous 18K gold, handcrafted to perfection',
+                                image: 'https://images.unsplash.com/photo-1605100804763-247f67b3557e?w=800&q=80'
+                            },
+                            {
+                                id: 2,
+                                name: 'Ceylon Moonstone Necklace',
+                                category: 'necklace',
+                                metal: 'White Gold',
+                                price: 1450,
+                                description: 'Authentic Ceylon moonstone with ethereal glow, beautifully set in white gold',
+                                image: 'https://images.unsplash.com/photo-1599643478518-a784e5dc4c8f?w=800&q=80'
+                            },
+                            {
+                                id: 3,
+                                name: 'Ruby Sunset Earrings',
+                                category: 'earring',
+                                metal: 'Platinum',
+                                price: 1350,
+                                description: 'Stunning Burmese rubies paired with platinum for timeless elegance',
+                                image: 'https://images.unsplash.com/photo-1535632066927-ab7c9ab60908?w=800&q=80'
+                            }
+                        ].map((item, index) => (
+                            <motion.div
+                                key={item.id}
+                                initial={{ opacity: 0, y: 30 }}
+                                whileInView={{ opacity: 1, y: 0 }}
+                                viewport={{ once: true }}
+                                transition={{ duration: 0.5, delay: index * 0.1 }}
+                                className="rounded-lg overflow-hidden"
+                                style={{
+                                    background: 'rgba(30, 41, 59, 0.5)',
+                                    backdropFilter: 'blur(10px)',
+                                    border: '1px solid rgba(251, 191, 36, 0.1)',
+                                    boxShadow: '0 8px 32px rgba(0, 0, 0, 0.3)'
+                                }}
+                            >
+                                <motion.div
+                                    className="aspect-square overflow-hidden relative"
+                                    whileHover={{ scale: 1.05 }}
+                                    transition={{ duration: 0.3 }}
+                                    style={{ background: 'rgba(0, 0, 0, 0.4)' }}
+                                >
+                                    <img
+                                        src={item.image}
+                                        alt={item.name}
+                                        className="w-full h-full object-cover"
+                                        style={{ objectPosition: 'center' }}
+                                    />
+                                    <div className="absolute inset-0 bg-gradient-to-t from-black/60 to-transparent" />
+                                </motion.div>
+                                <div className="p-8">
+                                    <h3 className="mb-2 text-white" style={{ fontSize: '1.5rem', fontWeight: '600' }}>{item.name}</h3>
+                                    <p className="mb-3" style={{ fontSize: '0.95rem', lineHeight: '1.6', color: 'rgba(255, 255, 255, 0.7)' }}>
+                                        {item.description}
+                                    </p>
+                                    <p style={{
+                                        fontSize: '0.875rem',
+                                        color: 'rgba(251, 191, 36, 0.8)',
+                                        marginBottom: '0.75rem',
+                                        fontWeight: '500'
+                                    }}>
+                                        {item.metal}
+                                    </p>
+                                    <p className="mb-4" style={{
+                                        background: 'linear-gradient(to right, rgb(251, 191, 36), rgb(202, 138, 4))',
+                                        WebkitBackgroundClip: 'text',
+                                        WebkitTextFillColor: 'transparent',
+                                        backgroundClip: 'text',
+                                        fontWeight: '700',
+                                        fontSize: '1.5rem'
+                                    }}>
+                                        ${item.price}
+                                    </p>
+                                    <motion.button
+                                        whileHover={{ scale: 1.05 }}
+                                        whileTap={{ scale: 0.95 }}
+                                        onClick={() => {
+                                            addToCart({
+                                                id: `collection-${item.id}-${Date.now()}`,
+                                                category: item.category,
+                                                metal: item.metal,
+                                                metalPrice: item.price,
+                                                size: 'M',
+                                                estimatedPrice: item.price
+                                            });
+                                            setToastMessage('Item added to cart!');
+                                            setShowToast(true);
+                                        }}
+                                        style={{
+                                            width: '100%',
+                                            padding: '12px 24px',
+                                            background: 'linear-gradient(to right, rgb(251, 191, 36), rgb(202, 138, 4))',
+                                            color: 'black',
+                                            fontWeight: '600',
+                                            borderRadius: '8px',
+                                            border: 'none',
+                                            cursor: 'pointer',
+                                            transition: 'all 0.3s ease',
+                                            boxShadow: '0 4px 12px rgba(251, 191, 36, 0.3)'
+                                        }}
+                                    >
+                                        Add to Cart
+                                    </motion.button>
+                                </div>
+                            </motion.div>
+                        ))}
+                    </div>
+                </div>
+            </div>
+
+            {/* Footer */}
+            <Footer />
+        </div>
+    );
+}
